backend/pyproject.toml
[project]
name = "finance-manager-backend"
version = "0.1.0"
description = "Backend for Nivetek Finance Manager"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.109.0",
    "uvicorn>=0.27.0",
    "sqlmodel>=0.0.14",
    "asyncpg>=0.29.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt==3.2.2",
    "python-multipart>=0.0.6",
    "pydantic-settings>=2.1.0",
    "alembic>=1.13.1",
    "python-dateutil>=2.8.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
